Hamid

+++ strore.py
@@ -70,8 +70,8 @@ class Store:
                 
                 if name_store in all_store.keys():
                     self.curently_path_store = all_store[name_store]
-                    self.store_sell_data_path = f'/Users/mohamad/Documents/vscode/matlab/jsons_file/sell/sell_{all_store[name_store]}'
-                    
+                    self.store_sell_data_path = f'/Users/mohamad/Documents/vscode/matlab/jsons_file/sell/sell_{name_store}.json'
+                    print(self.store_sell_data_path)
                     return 1
                 else:
                     return 0
@@ -161,6 +161,8 @@ class Store:
                         else:
                             print("the unvalid input.")
                             continue
+                else:
+                    print('product not exist')
     
     def show_stores(self):
         all_store = {}
@@ -188,11 +190,11 @@ class Store:
                     continue
                 
     def show_store_product(self):
-        if not self.check_exist_file(self.curently_path_store):
+        if not self.check_exist_file(self.curently_path_depot):
             print("not exist any prouduct. go to main")
             return 0
         else:
-            # try:
+            try:
                 object_store = {}
                 with open(self.curently_path_store,'r') as file:
                     object_store = json.load(file)
@@ -202,10 +204,12 @@ class Store:
                         count += 1
                     self.data_product_curently_store=object_store
                     return 1
-            # except:
-            #     print(self.curently_path_store)
-            #     print("not found store.")
+            except:
+                
+                print("not found store.")
                     
+    
+            
         
     def Buy_product(self,inp_prouduct : dict):
         select_product = input("Please enter the name product:")
@@ -218,41 +222,60 @@ class Store:
             else:
                 print("bad input.")
         else:
-            number_product = input("please enter number of product: ")
-            check = self.check_stock(inp_prouduct,number_product,select_product)
-            if not check:
-                print("This value is not available,Try Again")
-                self.Buy_product(inp_prouduct)
-            else:
-                inp_prouduct[select_product][0] = check
-                inp = input("going shopping cart?(1 = yes , 0 = no)")
-                if inp == '1':
-                    self.shopping_cart()
-                    self.data_product_curently_store = inp_prouduct
-                elif inp == '0':
-                    self.Buy_product()
+            try:
+                number_product = int(input("please enter number of product: "))
+                username = input("please enter username: ")
+                check = self.check_save_stock(inp_prouduct,number_product,select_product,username)
+                if not check:
+                    print(check)
+                    print("This value is not available,Try Again")
+                    self.Buy_product(inp_prouduct)
                 else:
-                    print('bad input,fuck you.\n shopping cart.....')
-                    self.shopping_cart()  
+                    inp_prouduct[select_product][0] = check
+                    inp = input("going shopping cart?(1 = yes , 0 = no)")
+                    if inp == '1':
+                        self.shopping_cart()
+                        self.data_product_curently_store = inp_prouduct
+                    elif inp == '0':
+                        count = 1
+                        for i ,j in self.data_product_curently_store.items():
+                            print(f"{count} - {i} : price = {j[1]} : stock = {j[0]} : Volume = {j[2]} :Description  = {j[3]}")
+                            count += 1
+                        self.Buy_product(inp_prouduct)
+                    else:
+                        print('bad input,fuck you.\n shopping cart.....')
+                        self.shopping_cart() 
+            except:
+                print("erooor") 
                       
-    def check_save_stock(self,inp_product:dict,number,name):
-        if number > inp_product[name][0]:
-            return 0
-        else:
-            date = str(persian_time.ShamsiDateTime())
-            now = datetime.now()
-            current_time = str(now.strftime("%H:%M:%S"))
+    def check_save_stock(self, inp_product: dict, number: int, name: str, username: str) -> int:
+        available_stock = int(inp_product[name][0])
+
+        if number <= available_stock:
+            try:
+                date = str(persian_time.ShamsiDateTime())
+                now = datetime.now()
+                current_time = now.strftime("%H:%M:%S")
+
+                if name in self.shopping_cart_dict:
+                    self.shopping_cart_dict[name][0] += number
+                else:
+                    self.shopping_cart_dict[name] = [[username, number, inp_product[name][1], date, current_time]]
+
+                return available_stock - number
+            except:
+                return 0
+        
+        return 0
             
-            self.shopping_cart_dict[name] = [number,inp_product[name][1],date,current_time]
-                
-            return  int(inp_product[name][0])-number
         
     
     def shopping_cart(self):
         inp = input("1 = Payment\n2 = Menu")          
         n = 1
-        for i ,j in self.shopping_cart.items():
-            print(f'{n} - {i} | {j[0]} | {j[0]*j[1]}')
+        
+        for i ,j in self.shopping_cart_dict.items():
+            print(f'{n} - {i} | {j[0][1]} | {j[0][1]*j[0][2]} | {j[0][0]}')
             n+=1
         if inp == '1':
             self.payment()
@@ -262,20 +285,40 @@ class Store:
             print("bad input.")
             
     def save_all_data(self):
-        try:
-            with open(self.curently_path_store,'w') as file:
-                json.dump(self.data_product_curently_store,file,indent=4)
+        
+            if not self.check_exist_file(self.curently_path_store):
+                with open(self.curently_path_store,'w') as file:
+                    json.dump(self.data_product_curently_store,file,indent=4)
+            else:
+                with open(self.curently_path_store,'w') as file:
+                    json.dump(self.data_product_curently_store,file,indent=4)
+                
             if not self.check_exist_file(self.store_sell_data_path):
                 with open(self.store_sell_data_path,'w') as file:
                     json.dump(self.shopping_cart_dict,file,indent=4)
             else:
                 data = {}
                 with open(self.store_sell_data_path,'r') as file:
-                    data = json.load(file)
-                        
-
-        except:
-            return 0 
+                    data:dict = json.load(file) 
+                for i ,j in self.shopping_cart_dict.items():
+                    if i in data.keys():
+                        data[i].append(j.pop())
+                    else:
+                        data[i] = j.pop()
+                
+                with open(self.store_sell_data_path,'w') as file:
+                    json.dump(data,file,indent=4)
+                         
+    def show_sell_store(self):
+        if not self.check_exist_file(self.store_sell_data_path):
+            print("not exist.")
+        else:
+            data = {}
+            with open(self.store_sell_data_path,'r') as file:
+                data = json.load(file)
+            for i ,j in data.items():
+                for k in j:
+                    print(f"{i} = {k[0]} , {k[1]} , {k[2]} , {k[3]} , {k[4]} ")    
     def delete_product(self,name:str):
         if not self.check_exist_file(self.curently_path_store):
             print("not exist.")
@@ -307,12 +350,6 @@ class Store:
                 else:
                     print("not store.")
         
-        
-        
-        
-        
-        
-    
                     
     def payment(self):
         inp = input("are you sure for payment ?? (Yes = 1 , No = 0)")
@@ -324,16 +361,16 @@ class Store:
             else:
                 pass
         elif inp == '0':
+            # self.data_product_curently_store = self.da
             print("menu...")
             pass
           
     def authorization_Admin(self):
-        while(1):
-            
+        while(1): 
             with open("/Users/mohamad/Documents/vscode/matlab/jsons_file/admin/Admins_user.json",'r') as users_seller:
                 username = input("please enter user name:(Exit = 0)") 
                 if username == '0':
-                    break
+                    return 0
                 password = input("please enter password: ")
                 users_data = json.load(users_seller)
                 if username in users_data.keys():
@@ -346,51 +383,65 @@ class Store:
                             continue   
                     return 0                          
     def menu_admin(self):
-        while(1):
-            self.authorization_Admin()
-            inp1 = input('Create or update product = 1 , Delete product = 2 , Delete store = 3 , Create store = 4  , Exit = 0')
-            
-            
-            if inp1 == '1':
-                inp = input("Add product = 1 , Update product = 2 , Exit = 0: ")
-                self.show_stores()
-                if inp == '1':
-                    name = input('please enter name: ')
-                    number = input('please enter number: ')
-                    price = input('please enter price: ')
-                    volume = input('please enter volume: ')
-                    des = input('please enter Description: ')
-                    self.add_object_store(name,int(number),int(price),int(volume),des)
-                    print("added")
-                elif inp == '2':
-                    self.show_store_product()
-                    name = input('please enter name: ')
-                    self.update_object_store(name)
-                    print("updated.")
-                elif inp == '0':
-                    pass
-                else:
-                    pass
-            elif inp1 == '2':
-                self.show_stores()
-                self.show_store_product()
-                name = input('please enter name: ')
-                self.delete_product(name)
+        check = self.authorization_Admin()
+        if check == 0:
+            print("cant open")
+            return 0 
+        else:  
+            while(1):
+                inp1 = input('Create or update product = 1 , Delete product = 2 , Delete store = 3 , Create store = 4 , Show_sell = 5 Exit = 0')
                 
-            elif inp1 == '3':
-                self.show_stores()
-                name = input('please enter name: ')
-                self.delete_store(name)
-            elif inp1 == '4':
-                name = input('please enter name: ')
-                self.create_store(name)
-                print("created") 
-            elif inp1 == '0':
-                break 
-            else:
-                print("by.") 
+                
+                if inp1 == '1':
+                    inp = input("Add product = 1 , Update product = 2 , Exit = 0: ")
+                    self.show_stores()
+                    if inp == '1':
+                        name = input('please enter name: ')
+                        number = input('please enter number: ')
+                        price = input('please enter price: ')
+                        volume = input('please enter volume: ')
+                        des = input('please enter Description: ')
+                        
+                        self.add_object_store(name,int(number),int(price),int(volume),des)
+                        print("added")
+                    elif inp == '2':
+                        self.show_store_product()
+                        name = input('please enter name: ')
+                        self.update_object_store(name)
+                        print("updated.")
+                    elif inp == '0':
+                        pass
+                    
+                    else:
+                        pass
+                elif inp1 == '2':
+                    self.show_stores()
+                    self.show_store_product()
+                    inp = input("are you sure?(0 = no , 1 = yes)")
+                    if inp == "0":
+                        return 0
+                    elif inp == "1":
+                        self.delete_product(name)
+                    else:
+                        print("bad input")
+                    
+                elif inp1 == '3':
+                    self.show_stores()
+                    name = input('please enter name: ')
+                    self.delete_store(name)
+                elif inp1 == '4':
+                    name = input('please enter name: ')
+                    self.create_store(name)
+                    print("created") 
+                elif inp1 == '5':
+                    self.show_stores()
+                    self.show_sell_store()
+                elif inp1 == '0':
+                    break 
+                else:
+                    print("by.") 
     def menu(self)->None:
-            inp = input("1 = Show product\n2 = Shopping cart\n3 = Menu\n4 = Exit :")
+            inp = input("1 = Show product\n2 = Shopping cart\n3 = Menu\n :")
             if inp == '1':
                 while(1):
                     if self.show_stores():
@@ -400,24 +451,33 @@ class Store:
                 if not self.show_store_product():
                     pass
                 else:
+                    f = True
                     self.Buy_product(self.data_product_curently_store)
                     while(1):
                         inp = input("Buy again = 1 , Continue = 2 :")
                         if inp == '1':
-                            self.show_store_product()
                             self.Buy_product(self.data_product_curently_store)
                         elif inp == '2':
+                            f = False
                             break
                         else:
                             print("bad input, try again.")
-                    self.payment()
-                    self.save_all_data()
+                    if f:
+                        self.payment()
+                        self.save_all_data()
+                    else:
+                        pass
             elif inp == '2':
-                self.shopping_cart()
+                inp = input("1 = Delete shoping cart\n2 = See the shoping cart : ")
+                if inp == '1':
+                    self.shopping_cart_dict = {}
+                    print("deleted.")
+                elif inp == '2':
+                    self.shopping_cart()
+                else:
+                    print("bad inputs.")
             elif inp == '3':
                 pass
-            elif inp == '4':
-                exit()
             else:
                 print("Bad input.Menu......")
                 pass
